wip of refactor
