+++ cmd/app/baseCmd.go
@@ -1,0 +1,13 @@
+package app
+
+import "github.com/spf13/cobra"
+
+const (
+	ParameterAccount  = "account"
+	ParameterProvider = "provider"
+)
+
+var appCommand = &cobra.Command{
+	Use:   "app",
+	Short: "Get and manage applications",
+}
