+++ cmd/config/baseCmd.go
@@ -1,0 +1,8 @@
+package config
+
+import "github.com/spf13/cobra"
+
+var configBaseCommand = &cobra.Command{
+	Use:   "config",
+	Short: "View and get cli configuration",
+}
