+++ go.mod
@@ -4,6 +4,7 @@ go 1.16
 
 require (
 	github.com/armory-io/go-cloud-service v0.8.0-79558ea
+	github.com/golang/groupcache v0.0.0-20210331224755-41bb18bfe9da // indirect
 	github.com/golang/protobuf v1.5.2
 	github.com/hashicorp/go-getter v1.5.6
 	github.com/juju/ansiterm v0.0.0-20210706145210-9283cdf370b5
@@ -14,13 +15,12 @@ require (
 	github.com/spf13/pflag v1.0.5
 	github.com/stretchr/objx v0.2.0 // indirect
 	github.com/stretchr/testify v1.7.0
-	golang.org/x/lint v0.0.0-20210508222113-6edffad5e616 // indirect
+	golang.org/x/crypto v0.0.0-20210220033148-5ea612d1eb83 // indirect
+	golang.org/x/time v0.0.0-20210723032227-1f47c861a9ac // indirect
 	golang.org/x/tools v0.1.5 // indirect
 	google.golang.org/genproto v0.0.0-20210602131652-f16073e35f0c
 	google.golang.org/grpc v1.38.0
 	google.golang.org/protobuf v1.26.0
 	gopkg.in/yaml.v2 v2.4.0
-	k8s.io/api v0.22.0
 	k8s.io/apimachinery v0.22.0
-	k8s.io/client-go v0.22.0
 )
